perf: avoid modification on abstract model
- instead of inheriting abstract model, better inherit sale order directly
