+++ addons/abp_customer_catalogue/models/__init__.py
@@ -1,6 +1,5 @@
 from . import customer_catalogue
 # from . import pos_session
-from . import product_catalog_mixin
 from . import product_template
 from . import res_partner
 from . import sale_order_line
