+++ addons/abp_customer_catalogue/models/sale_order.py
@@ -94,3 +94,23 @@ class SaleOrder(models.Model):
         # Send the email
         mail_id.send()
         
+        
+    def _get_product_catalog_additional_domain(self):
+        order = self
+        if not order.show_all_product:
+            customer_catalogue = self.env['customer.catalogue'].search([
+                ('partner_id', '=', order.partner_id.id)
+            ])
+            product_ids = customer_catalogue.mapped('product_id').mapped('id')
+            additional_domain = [('id', 'in', product_ids)]
+            return additional_domain
+            
+            
+    def _get_product_catalog_domain(self):
+        result = super()._get_product_catalog_domain()
+        
+        additional_domain = self._get_product_catalog_additional_domain()
+        if additional_domain:
+            result += additional_domain
+            
+        return result
